Split off Widgets::BaseMixin from the widgets
