+++ src/lib/y2configuration_management/widgets.rb
@@ -17,6 +17,7 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
+require "y2configuration_management/widgets/base_mixin"
 require "y2configuration_management/widgets/boolean"
 require "y2configuration_management/widgets/collection"
 require "y2configuration_management/widgets/group"
