+++ test/y2configuration_management/widgets/boolean_test.rb
@@ -20,11 +20,10 @@
 require_relative "../../spec_helper"
 require "y2configuration_management/widgets/boolean"
 require "y2configuration_management/salt/form"
-require "y2configuration_management/salt/form_controller"
 require "cwm/rspec"
 
 describe Y2ConfigurationManagement::Widgets::Boolean do
-  subject(:widget) { described_class.new(spec, controller) }
+  subject(:widget) { described_class.new(spec) }
 
   include_examples "CWM::CheckBox"
 
@@ -33,11 +32,10 @@ describe Y2ConfigurationManagement::Widgets::Boolean do
   end
   let(:spec) { form_spec.find_element_by(locator: locator) }
   let(:locator) { ".root.person.wants_newsletter" }
-  let(:controller) { instance_double(Y2ConfigurationManagement::Salt::FormController) }
 
   describe ".new" do
     it "instantiates a new widget according to the spec" do
-      widget = described_class.new(spec, controller)
+      widget = described_class.new(spec)
       expect(widget.locator).to eq(locator)
       expect(widget.default).to eq(true)
     end
